refactor: rename file to match lsp
